+++ vm/string.c
@@ -111,9 +111,9 @@ OBJ TrString_cmp(VM, OBJ self, OBJ other) {
 }
 
 OBJ TrString_substring(VM, OBJ self, OBJ start, OBJ len) {
-  int s = TR_FIX2INT(start);
-  int l = TR_FIX2INT(len);
-  if (s < 0 || (s+l) > (int)TR_STR_LEN(self)) return TR_NIL;
+  intptr_t s = TR_FIX2INT(start);
+  intptr_t l = TR_FIX2INT(len);
+  if (s < 0 || (s+l) > (intptr_t)TR_STR_LEN(self)) return TR_NIL;
   return TrString_new(vm, TR_STR_PTR(self)+s, l);
 }
